relocate queries from annual tab
